+++ src/main/java/com/nbrichau/vanillaextension/slabs/GrassBlockSlab.java
@@ -72,7 +72,7 @@ public class GrassBlockSlab extends SlabBlock implements IGrowable {
 				if (blockstate2.isAir()) {
 					BlockState blockstate1;
 					if (rand.nextInt(8) == 0) {
-						List<ConfiguredFeature<?, ?>> list = worldIn.getBiome(blockpos1).func_242440_e().func_242496_b();
+						List<ConfiguredFeature<?, ?>> list = worldIn.getBiome(blockpos1).getGenerationSettings().getFlowerFeatures();
 						if (list.isEmpty()) {
 							continue;
 						}