+++ src/main/java/com/nbrichau/vanillaextension/trapdoors/ConcretePowderTrapdoor.java
@@ -197,7 +197,7 @@ public class ConcretePowderTrapdoor extends ConcretePowderBlock implements IWate
 		for (Direction direction : Direction.values()) {
 			BlockState blockstate = reader.getBlockState(blockpos$mutable);
 			if (direction != Direction.DOWN || causesSolidify(blockstate)) {
-				blockpos$mutable.func_239622_a_(pos, direction);
+				blockpos$mutable.setAndMove(pos, direction);
 				blockstate = reader.getBlockState(blockpos$mutable);
 				if (causesSolidify(blockstate) && !blockstate.isSolidSide(reader, pos, direction.getOpposite())) {
 					flag = true;